fix: change types in entities to correspoding MySQL types in java

+++ src/main/java/zlagoda/server/company/entity/CustomerCard.java
@@ -14,5 +14,5 @@ public class CustomerCard {
     private String city;
     private String street;
     private String zipCode;
-    private Long percent;
+    private int percent;
 }
