+++ src/main/java/zlagoda/server/company/entity/Employee.java
@@ -1,5 +1,6 @@
 package zlagoda.server.company.entity;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 import lombok.Data;
@@ -18,7 +19,7 @@ public class Employee {
     private String name;
     private String patronymic;
     private Role role;
-    private long salary;
+    private BigDecimal salary;
     private Date birthdate;
     private Date startDate;
     private String phoneNumber;
